BAEL-5323: merge sources into existing persistence module
