+++ persistence-modules/spring-data-jdbc/src/main/java/com/baeldung/springmultipledatasources/Application.java
@@ -1,0 +1,15 @@
+package com.baeldung.springmultipledatasources;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.core.env.AbstractEnvironment;
+
+@SpringBootApplication
+public class Application {
+
+    public static void main(String[] args) {
+        System.setProperty(AbstractEnvironment.ACTIVE_PROFILES_PROPERTY_NAME, "multipledatasources");
+        SpringApplication.run(Application.class, args);
+    }
+
+}
