+++ persistence-modules/spring-data-jdbc/src/main/java/com/baeldung/springmultipledatasources/todos/TodoRepository.java
@@ -1,4 +1,4 @@
-package com.baeldung.spring.datasources.todos;
+package com.baeldung.springmultipledatasources.todos;
 
 import org.springframework.data.jpa.repository.JpaRepository;
