+++ persistence-modules/spring-data-jdbc/src/main/resources/application-multipledatasources.properties
@@ -1,0 +1,9 @@
+spring.jpa.hibernate.ddl-auto=update
+spring.datasource.todos.url=jdbc:h2:mem:todos
+spring.datasource.todos.username=sa
+spring.datasource.todos.password=null
+spring.datasource.todos.driverClassName=org.h2.Driver
+spring.datasource.topics.url=jdbc:h2:mem:topics
+spring.datasource.topics.username=sa
+spring.datasource.topics.password=null
+spring.datasource.topics.driverClassName=org.h2.Driver
