+++ persistence-modules/spring-data-jdbc/src/test/java/com/baeldung/springmultipledatasources/MultipleDatasourcesIntegrationTest.java
@@ -1,17 +1,19 @@
-package com.baeldung.spring.datasources;
+package com.baeldung.springmultipledatasources;
 
-import com.baeldung.spring.datasources.todos.Todo;
-import com.baeldung.spring.datasources.todos.TodoRepository;
-import com.baeldung.spring.datasources.topics.Topic;
-import com.baeldung.spring.datasources.topics.TopicRepository;
+import com.baeldung.springmultipledatasources.todos.Todo;
+import com.baeldung.springmultipledatasources.todos.TodoRepository;
+import com.baeldung.springmultipledatasources.topics.Topic;
+import com.baeldung.springmultipledatasources.topics.TopicRepository;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
+import org.springframework.test.context.ActiveProfiles;
 
 import java.util.Optional;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
+@ActiveProfiles("multipledatasources")
 @DataJpaTest // no test database!
 class MultipleDatasourcesIntegrationTest {
