+++ spring-boot-modules/pom.xml
@@ -52,7 +52,6 @@
         <module>spring-boot-libraries</module>
         <module>spring-boot-libraries-2</module>
         <module>spring-boot-logging-log4j2</module>
-        <module>spring-boot-multiple-datasources</module>
         <module>spring-boot-mvc</module>
         <module>spring-boot-mvc-2</module>
         <module>spring-boot-mvc-3</module>
