--- spring-boot-modules/spring-boot-multiple-datasources/.gitignore
@@ -1,1 +1,0 @@
-.local-db
