--- spring-boot-modules/spring-boot-multiple-datasources/src/main/java/com/baeldung/spring/datasources/MultipleDatasourcesApplication.java
@@ -1,13 +1,0 @@
-package com.baeldung.spring.datasources;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class MultipleDatasourcesApplication {
-
-    public static void main(String[] args) {
-        SpringApplication.run(MultipleDatasourcesApplication.class, args);
-    }
-
-}
